deepin: net: txgbe: fix CONFIG_DCB=N compile error
wangxun inclusion
category: bugfix

fix set CONFIG_DCB=N, related txgbe dcb module compile error.
