+++ drivers/net/ethernet/wangxun/txgbe/txgbe_fcoe.c
@@ -935,6 +935,10 @@ int txgbe_fcoe_get_wwn(struct net_device *netdev, u64 *wwn, int type)
  */
 u8 txgbe_fcoe_get_tc(struct txgbe_adapter *adapter)
 {
+#if IS_ENABLED(CONFIG_DCB)
 	return netdev_get_prio_tc_map(adapter->netdev, adapter->fcoe.up);
+#else
+	return 0;
+#endif /* CONFIG_DCB */
 }
 #endif /* CONFIG_FCOE */
